get posts

+++ Models/Tables/Post.cs
@@ -13,6 +13,6 @@ public class Post
     public required ContentType ContentType { get; set; }
     public required string Content { get; set; }
     public string? Text { get; set; }
-    public required int TotalLikes { get; set; }
-    public required int TotalComments { get; set; }
+    public required long TotalLikes { get; set; }
+    public required long TotalComments { get; set; }
 }
